Adding mirrors top connect reflection to something real without going through Java.
git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25334 5e8d7ff9-d8ef-0310-90f0-a4852d11357a
